[feat] api crud pour la gestion des utilisateurs

+++ src/main/java/com/univ/cardgame/entity/UserEntity.java
@@ -2,6 +2,7 @@
 package com.univ.cardgame.entity;
 
 import com.univ.cardgame.constante.RoleEnum;
+import lombok.AllArgsConstructor;
 import lombok.Data;
 
 import javax.persistence.*;
@@ -12,7 +13,7 @@ import java.util.Collection;
 public class UserEntity {
     @Id
     @GeneratedValue(strategy = GenerationType.AUTO)
-    private Long id;
+    private Integer id;
 
     @Column(unique = true, nullable = false)
     private String login;
